+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2026-08-08
+
+- **@poc-plattform-kit/api** `0.15.0` → `0.16.0` (minor)
+- **@poc-plattform-kit/web** `0.11.0` → `0.12.0` (minor)
+- **@poc-plattform-kit/api-client** `0.8.0` → `0.9.0` (minor)
+- **@poc-plattform-kit/pillar-permissions** `0.2.0` → `0.3.0` (minor)
+- **@poc-plattform-kit/pillar-tenant** `0.3.0` → `0.4.0` (minor)
 ## 2026-08-07
 
 - **@poc-plattform-kit/api** `0.14.0` → `0.15.0` (minor)